% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_split_rules.R
\name{extract_split_rules}
\alias{extract_split_rules}
\title{Get split conditions for terminal nodes for all trees in model.}
\usage{
extract_split_rules(h2o_trees_in_R)
}
\arguments{
\item{h2o_trees_in_R}{\code{list} output from \code{h2o_tree_model_to_R}
function.}
}
\value{
\code{list} containing split conditions to reach terminal nodes for
        each tree.
}
\description{
This function takes the output from \code{h2o_tree_model_to_R} and for each
tree in the model determines the split conditions required to reach the
terminal nodes.
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

prostate.hex$RACE <- as.factor(prostate.hex$RACE)

prostate.hex$DPROS <- as.factor(prostate.hex$DPROS)

prostate.gbm = h2o.gbm(x = c("AGE", "RACE", "DPROS", "DCAPS", "PSA", "VOL", "GLEASON"),
                       y = "CAPSULE",
                       training_frame = prostate.hex,
                       ntrees = 5,
                       max_depth = 5,
                       learn_rate = 0.1)

h2o_trees <- h2o_tree_model_to_R(h2o_model = prostate.gbm,
                                 mojo_output_path = '/Users/UserName/Desktop',
                                 gv_output_path = '/Users/UserName/Desktop',
                                 model_ini_overwrite = TRUE,
                                 h2o_jar_file = 'h2o.jar')

h2o_trees_terminal_node_rules <- extract_split_rules(h2o_trees)

}
