% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_tree_model_to_R.R
\name{h2o_tree_model_to_R}
\alias{h2o_tree_model_to_R}
\title{Convert h2o tree .gv file to tabular structure.}
\usage{
h2o_tree_model_to_R(h2o_model, mojo_output_path, gv_output_path,
  model_ini_overwrite = TRUE, h2o_jar_file)
}
\arguments{
\item{h2o_model}{gbm or drf h2o model.}

\item{mojo_output_path}{directory to export h2o_model's MOJO .zip to.}

\item{gv_output_path}{directory to output .gv file for each tree in model.}

\item{model_ini_overwrite}{\code{logical} when extracting the model.ini
from the MOJO .zip, should an existing file be overwritten? Default
= \code{TRUE}.}

\item{h2o_jar_file}{h2o.jar file, including path.}
}
\value{
\code{data.frame} containing tree structure.
}
\description{
Reads a single .gv file containing a tree from a h2o gbm or drf and parses it
to a tabular structure.
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

prostate.gbm = h2o.gbm(x = c("AGE", "RACE", "DPROS", "DCAPS", "PSA", "VOL", "GLEASON"),
                       y = "CAPSULE",
                       training_frame = prostate.hex,
                       ntrees = 5,
                       max_depth = 5,
                       learn_rate = 0.1)

h2o_trees <- h2o_tree_model_to_R(h2o_model = prostate.gbm,
                                 mojo_output_path = '/Users/UserName/Desktop',
                                 gv_output_path = '/Users/UserName/Desktop',
                                 model_ini_overwrite = TRUE,
                                 h2o_jar_file = 'h2o.jar')

}
