% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo_trees_to_gvs.R
\name{mojo_trees_to_gvs}
\alias{mojo_trees_to_gvs}
\title{Export all trees from a h2o mojo (.zip) to .gv files}
\usage{
mojo_trees_to_gvs(h2o_jar, mojo_zip, gv_output_dir)
}
\arguments{
\item{h2o_jar}{h2o.jar file, including path.}

\item{mojo_zip}{h2o tree based mojo zip file, including path.}

\item{gv_output_dir}{directory to output .gv files to.}
}
\value{
Outputs trees from mojo zip to .gv files.
}
\description{
Calls call_PrintMojo to output each tree to .gv file in the input directory
}
