% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_h2o_frame.R
\name{sample_h2o_frame}
\alias{sample_h2o_frame}
\title{Sample a H2OFrame.}
\usage{
sample_h2o_frame(h2o_frame, seed = 1, x = NULL, size, replace = FALSE,
  prob = NULL)
}
\arguments{
\item{h2o_frame}{\code{H2OFrame} to sample by rows.}

\item{seed}{seed for random number generation.}

\item{x}{Either a vector of one or more elements from which to choose, or
a positive integer. Passed to \code{x} in \code{sample}. Cannot
have negative elements or elements larger than nrow \code{H2OFrame}.}

\item{replace}{\code{logical} Should sampling be with replacement? Passed
to \code{replace} in \code{sample}.}

\item{prob}{A vector of probability weights for obtaining the elements of
the vector being sampled. Passed to \code{prob} in \code{sample}.}
}
\value{
Sampled \code{H2OFrame}.
}
\description{
Sample a H2OFrame.
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

sample_h2o_frame(h2o_frame = prostate.hex, size = 100, replace = TRUE)

}
