% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_split_rules.R
\name{terminal_node_exprs}
\alias{terminal_node_exprs}
\title{Get split conditions for terminal nodes.}
\usage{
terminal_node_exprs(split_exprs)
}
\arguments{
\item{split_exprs}{\code{data.frame} output from get_split_expressions function.}
}
\value{
\code{data.frame} containing complete split conditions for terminal nodes.
}
\description{
This is a wrapper for get_terminal_node_directions to do input checking and return
results in a better form (\code{data.frame}).
}
\examples{
library(h2o)

h2o.init()

prostate.hex = h2o.uploadFile(path = system.file("extdata",
                                                 "prostate.csv",
                                                 package = "h2o"),
                              destination_frame = "prostate.hex")

prostate.hex$RACE <- as.factor(prostate.hex$RACE)

prostate.hex$DPROS <- as.factor(prostate.hex$DPROS)

prostate.gbm = h2o.gbm(x = c("AGE", "RACE", "DPROS", "DCAPS", "PSA", "VOL", "GLEASON"),
                       y = "CAPSULE",
                       training_frame = prostate.hex,
                       ntrees = 5,
                       max_depth = 5,
                       learn_rate = 0.1)

h2o_trees <- h2o_tree_model_to_R(h2o_model = prostate.gbm,
                                 mojo_output_path = '/Users/UserName/Desktop',
                                 gv_output_path = '/Users/UserName/Desktop',
                                 model_ini_overwrite = TRUE,
                                 h2o_jar_file = 'h2o.jar')

tree_1_split_expr <- get_split_expressions(h2o_trees[[1]])

tree_1_terminal_nodes_df <- terminal_node_exprs(df = tree_1_split_expr)

}
