% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_PrintMojo.R
\name{call_PrintMojo}
\alias{call_PrintMojo}
\title{Export single h2o tree from MOJO .zip to .gv file.}
\usage{
call_PrintMojo(h2o_jar, tree_no, mojo_zip, gv_output)
}
\arguments{
\item{h2o_jar}{h2o.jar file, including path.}

\item{tree_no}{tree number to export.}

\item{mojo_zip}{h2o tree based MOJO zip file, including path.}

\item{gv_output}{.gv file to output to, including path.}
}
\value{
Outputs tree from MOJO zip to specified .gv file.
}
\description{
Calls the h2o hex.genmodel.tools.PrintMojo function.
}
\examples{
call_PrintMojo('h2o.jar',
               4,
               'GBM_model_R_1488095800763_37.zip',
               'GBM_model_R_1488095800763_37_4.gv')

}
