% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojo_gv_to_table.R
\name{mojo_gv_to_table}
\alias{mojo_gv_to_table}
\title{Convert h2o tree .gv file to tabular structure}
\usage{
mojo_gv_to_table(gv_file)
}
\arguments{
\item{gv_file}{.gv file file including path to import.}
}
\value{
\code{data.frame} containing tree structure.
}
\description{
Reads a single .gv file containing a tree from a h2o gbm or drf and parses it
into a tabular structure
}
